Add constructor for DateTimeOffset
- Change nanos back to i64 since it has to hold a u32
- Have to pick between using .timestamp() or Duration arithmetic

+++ bolt-proto/src/value/date_time_zoned.rs
@@ -13,7 +13,7 @@ pub(crate) const SIGNATURE: u8 = 0x44;
 #[derive(Debug, Clone, Hash, Eq, PartialEq, Signature, Marker, Serialize, Deserialize)]
 pub struct DateTimeZoned {
     pub(crate) epoch_seconds: i64,
-    pub(crate) nanos: i32,
+    pub(crate) nanos: i64,
     pub(crate) zone_id: String,
 }
